trying to compile

+++ UPDATED_VERSION/contracts/utils/Withdrawable.sol
@@ -1,3 +1,5 @@
+//SPDX-License-Identifier: MIT
+
 pragma solidity ^0.6.6;
 
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
@@ -8,7 +10,7 @@ import "@openzeppelin/contracts/access/Ownable.sol";
     Ensures that any contract that inherits from this contract is able to
     withdraw funds that are accidentally received or stuck.
  */
- 
+
 contract Withdrawable is Ownable {
     using SafeERC20 for ERC20;
     address constant ETHER = address(0);
@@ -16,7 +18,7 @@ contract Withdrawable is Ownable {
     event LogWithdraw(
         address indexed _from,
         address indexed _assetAddress,
-        uint amount
+        uint256 amount
     );
 
     /**
@@ -24,7 +26,7 @@ contract Withdrawable is Ownable {
      * @param _assetAddress Asset to be withdrawn.
      */
     function withdraw(address _assetAddress) public onlyOwner {
-        uint assetBalance;
+        uint256 assetBalance;
         if (_assetAddress == ETHER) {
             address self = address(this); // workaround for a possible solidity bug
             assetBalance = self.balance;
